requirements.txt
cycler==0.11.0
feather-format==0.4.1
fonttools==4.33.3
kiwisolver==1.4.2
matplotlib==3.5.2
numpy==1.20.2
packaging==21.3
pandas==1.4.2
Pillow==8.2.0
pyarrow==8.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
jinja2<3.1.0